test: add unit test for search history across databases
